move utils to were it belongs
