+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './artifact-downloader.js'
 export * from './definitions.js'
+export * from './utils.js'
